tests/unit/test_user_endpoints.py
import pytest
from fastapi import status


def test_user_endpoints():
    # Placeholder for user endpoints unit tests
    assert True, "Implement user endpoints unit tests here."


def test_create_user(client, test_user_data):
    """Test user creation endpoint"""
    response = client.post("/api/v1/users", json=test_user_data)
    assert response.status_code == status.HTTP_200_OK
    assert "id" in response.json()
    assert response.json()["email"] == test_user_data["email"]


def test_create_duplicate_user(client, test_user_data):
    """Test creating user with duplicate email"""
    # Create first user
    client.post("/api/v1/users", json=test_user_data)

    # Try to create duplicate
    response = client.post("/api/v1/users", json=test_user_data)
    assert response.status_code == status.HTTP_400_BAD_REQUEST


def test_get_user(client, test_user_data, auth_headers):
    """Test getting user details"""
    # Create user first
    create_response = client.post("/api/v1/users", json=test_user_data)
    user_id = create_response.json()["id"]

    # Get user details
    response = client.get(f"/api/v1/users/{user_id}", headers=auth_headers)
    assert response.status_code == status.HTTP_200_OK
    assert response.json()["email"] == test_user_data["email"]


def test_get_nonexistent_user(client, auth_headers):
    """Test getting non-existent user"""
    response = client.get("/api/v1/users/999", headers=auth_headers)
    assert response.status_code == status.HTTP_404_NOT_FOUND


@pytest.fixture
def test_user_data():
    """Test user credentials"""
    return {"email": "test@example.com", "password": "TestPass123!"}
